+++ secrets/secrets.nix
@@ -23,4 +23,9 @@ in {
     inherit (users) enzime_phi;
     inherit (hosts) phi;
   };
+
+  "zshrc_tau.age".publicKeys = builtins.attrValues {
+    inherit (users) enzime_tau;
+    inherit (hosts) tau;
+  };
 }